Add files via upload
Got rid of the trails following the player dot along.  Each sprite can also now be assigned a color, also. I would like to add a static redraw method for this class too: Right now, redrawing is done whenever a key is pressed. When AI is introduced, there will be lots of events that can trigger a redraw.